fix float calculate

+++ web/js/buy/booking.js
@@ -87,7 +87,7 @@ $(document).ready(function () {
         });
 
         total = $.helper.round(total, 2);
-        var limit = parseFloat($('#buyLimit').val());
+        var limit = parseInt($('#buyLimit').val());
 
         $('#tongji .realprice').html(total);
         if (total >= limit) {
